requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
fake-useragent==1.5.1
pyspark==3.5.1
pandas==2.2.1
pyarrow==15.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3
tqdm==4.66.2
google-genai==1.10.0
dbt-core==1.7.10
dbt-postgres==1.7.10
